Rename AppDND and App
